fix(sem): `nimsuggest` crashing for ill-formed AST (#939)
## Summary

Fix `nimsuggest`, and other tooling that uses a custom structured
report hook, crashing when encountering ill-formed AST that has the
wrong number of nodes.

## Details

Many of the semantic analysis procedures first verify that the input
`PNode` has the right number of child nodes. This is done via the
`checkSonsLen` and `checkMinSonsLen` procedures, with the expectation
being that both raise an `ERecoverableError` (through `globalReport`)
when the check fails.

However, since reporting the error happens via `handleReport`, it's
possible that the error handling strategy is overridden with a mode 
(e.g., `doNothing`) that results in execution to continue without any
exception being raised. `nimsuggest` sets the error handling mode to
`doNothing`, which then resulted in `checkSonsLen` and
`checkMinSonsLen` not ending the active routine, usually leading to
`IndexDefect`s or other crashes.

As an interim solution, the `reportAndFail` procedure is introduced.
`reportAndForceRaise` works similar to similar to `handleReport`, but
it ensures that *at least* an exception is raised -- both
`checkSonsLen` and `checkMinSonsLen` use it to ensure that the calling
routine is exited.

---------

Co-authored-by: zerbina <100542850+zerbina@users.noreply.github.com>