`semFor` propagate iterator position error nodes (#932)
## Summary

`semFor`  propagates errors that occur in the iterator position. They
are
now properly reported and `semFor` produces errors as required.

## Details

Updated `semFor` to propagate error nodes in the iterator position, as
well as appropriately wrap the semantically analysed  `nkFor`  if an
error
is present.

Added `adSemForExpectedIterator` to `AstDiagKind` to capture the error
as a node instead of using legacy reports `localReport`.

---------

Co-authored-by: Saem Ghani <saemghani+github@gmail.com>