make the `in` magic work the same across backends
Summary
=======

Consistently skip range checks for the first operand of an `in` call
across all backends. The special behaviour was previously only enabled
for the C backend.

Details
=======

While the C code-generator turning an `in` call into a chain of
comparisons looks to have been intended as an optimization only, it was
at some point extended with skipping the range check, resulting in
different behaviour compared to the un-optimized version.

The optimization/lowering is now performed during the AST -> MIR step
(`mirgen`), for the following reasons:
- the range check has control-flow implications, and must thus be
  eliminated prior to control-flow dependent MIR passes
- a MIR-level lowering means that it works the same across all backends

While the optimization could have also be applied in `transf`, the MIR,
with its region construct, is a better fit. The lowering can also not
happen as a MIR pass, as it relies on the exact code shape, which is
not guaranteed to stay the same during MIR processing.

The conditions for when to apply the optimization are kept the same.
Two MIR code-size optimization are applied:
- literal values are not passed as region arguments, but are instead
  passed directly to the comparison operations (gets around an
  `mnkArg` + `mnkOpParam` node, and also means less work for `cgirgen`)
- range comparisons (`a <= x and x <= b`) use a `bitand` instead of a
  boolean `and` -- this gets around having to generate another `if`
  statement

For choosing the right `mLeX` magic, the `getMagicLeForType` procedure
is added to the `magicsys` module.

Finally, a test for the operator's special behaviour is added.