vm: convert `slurp` and `staticRead` to vmops (#947)
## Summary

Convert `slurp` and `staticRead` in the `system` module from magics to
VM callback operations.

## Details

Removal of the magic requires creating a conditional compilation step,
here `nimskullReworkStaticExec` conditional symbol was resued (see:
`condsym`). Two definitions of `slurp` and `staticRead` now exist, one
with the magic and one without. Without this, bootstrapping would no
longer work.

Also marked `slurp` as deprecated, `staticRead` is a far better name.
The reason for the deprecation is that `slurp` is just not as clear as
`staticRead`.

Removed `mSlurp` magic enum, and the `opcSlurp` VM opcode.