CI: use strict mode for compiling (#951)
## Summary

Extend `nimStrictMode` to error on unused imports, and enable
`nimStrictMode` when compiling the compiler and core tools during CI
runs.

The idea is to have this help with catching bugs and preventing
leftover debug code being merged, while not complicating local
development. Over time, `nimStrictMode` should promote more warnings
and hints to errors.

## Details

* treat `UnusedImport` warnings as an error when `nimStrictMode` is
  enabled
* remove the obsolete `runCI` koch command
* introduce the new `all-strict` koch command, which does the same
  as `all`, except that it enables strict mode
* when building the compiler and tools in strict mode, use an
  `--errorMax` of 3, so that compilation doesn't abort after the first
  error
* run `all-strict` instead of `all` in the CI

Due to a bug, the current csources compiler doesn't properly support
warnings/hints promoted to errors, so `nimStrictMode` is turned into a
no-op when `nimKochBootstrap` is defined (this define is set for the
first iteration of bootstrapping). This is a temporary workaround,
which can be removed after the next csources update.