fix(sigmatch): AST corruption of bracket calls (#961)
## Summary

Fix explicit bracket calls (`[](...)`) and bracket expressions in
generic procedures or templates causing AST corruption.

A concrete example where this caused problems is when a bracket call is
used as the argument to an `untyped` parameter of an overloaded
template where there exist overloads with typed parameters but none
of them matches.

## Details

### Problem description

For bracket calls, the overload set contains the `mArrGet` or `mArrPut`
magics -- these always match.

For `mArrGet`  and `mArrPut` candidates, `sigmatch.matches` assigns the
*original* call AST, which is a problem when the candidate is the one
picked in the end: `semResolvedCall` expects a call AST that it can
freely modify, but since the original AST is used, the modifications
show up in unintended places (such as on the AST for earlier processed
candidates).

Due to bracket expression (`nkBracketExpr`) rewriting performed for the
bodies of templates and generic procedures, bracket expressions within
those were also affected.

### The solution

In order to uphold the expectation that the immediate call AST of a
`TCandidate`  can be modified if the candidate represents a match,
change
`matches` to create a shallow copy of the call node for a `mArrGet`/
`mArrPut` match.

---------

Co-authored-by: Saem Ghani <saemghani+github@gmail.com>