refactor(lookups): simpler identifier error handling
Summary
=======

Simplify identifier error handling in `qualifiedLookUp`, removing a
duplication of how errors for potentially quoted identifier are
produced.

Details
=======

* replace the `isNotFound` test and subsequent `nkError` creation with
  directly using the `errNode` returned by `considerQuotedIdent`
* remove the now obsolete `errorExpectedIdentifier` procedure

There is a small difference in the "wrong node" selection now, but the
previous logic in `qualifiedLookUp` was incorrect (the erroneous node
is the `nkAccQuoted`, not its element), so this is an improvement in
error precision.