refactor(sem): remove `fixupTypeAfterEval`
Summary
=======

Remove the obsolete `fixupTypeAfterEval` procedure and all its usages.

Details
=======

`fixupTypeAfterEval` was originally required for ensuring the value AST
returned by the VM is properly typed, but with the introduction of
`vmcompilerserdes` (https://github.com/nim-works/nimskull/pull/242),
this became obsolete.

Remaining usages
----------------

Normal procedures procedures cannot return `untyped`, `typed`, or
`typeDesc` values, meaning that the `fixupTypeAfterEval` in
`evalAtCompileTime` was always a no-op.

For `static` expression, `typeDesc` values are possible, but since the
evaluated result coming out of `evalStaticExpr` is always a `nkType`
AST, for which `semExpr` is a no-op, `fixupTypeAfterEval` was also a
no-op there.