fix(sem): named param errors no longer crash (#959)
## Summary

Errors in named parameter call sites no longer cause the compiler to
crash.

## Details

A logic error in `semcall.maybeResemArgs` overwrote `results` with
incorrect data leading to a crash. This has been fixed.

In addition added some extra information to assertions that originally
led to the discovery of this error in order to assist debugging these
sorts of issues more quickly in the future.

A regression test was added to ensure errors in named parameters no
longer crash the compiler.

A few minor formatting and stylistic clean-ups were made in related
code.