fix(suggest): project parents non-project modules (#953)
## Summary

`suggest` correctly falls back to treating the project module as the
parent module for out of project files/modules.

## Details

Prior to this change, querying the parent module for a file that is not
the project or a dependent module in the the project (via `import` or
`include`), resulted in a returning a file index of `0`. This resulted
in the parent being treated as the config file (typically first to get
indexed).

After this change, parent module querying, via `parentModule`, where no
parent is found results in an `InvalidFileIdx`. Since this is used as
part of suggest we continue to make progress by assuming the project
module is the parent module and continue semantic analysis.

As part of this change a regression test was added to ensure this
continues to work.

---------

Co-authored-by: Saem Ghani <saemghani+github@gmail.com>