fix(sigmatch): built-in type-class constraint issue (#976)
## Summary

Fix overload resolution failing for procedures with signatures like
`proc p[A, B: <built-in-type-class>](a: A, b: B)` where the types used
as the arguments are not equal.

## Details

For type relation (`typeRel`) analysis between a type and a generic
type parameter constraint, binding types to type-classes is disabled
(the `trDontBind` flag is passed), but the logic for
`tyBuiltInTypeClass` ignored the request and always bound the actual
type to the type-class.

Since in `[A, B: <built-in-type-class>]`, the `tyBuiltInTypeClass`
`PType` instance used as the constraints is the same, once `A` was
bound, `B` effectively acted as being constrained to `A`, which caused
a type relation failure when the type passed to `B` is not equal to
that passed to `A`.

The `typeRel` logic for `tyBuiltInTypeClass` now respects the
`trDontBind` flag, fixing the issue. However, not all built-in type-
classes use the `tyBuiltInTypeClass` type at the moment (fixing that is
outside the scope of this commit), so for those that don't, the issue
persists.

---------

Co-authored-by: Saem Ghani <saemghani+github@gmail.com>