semtypes: fix the bind-once behaviour
Fix the pre-existing issue of type-classes of built-in types not having
"bind once" behaviour as specified in the manual.

Type-classes of built-in generic types that have a symbol (i.e.,
those defined in `system`) use the symbols from the type-classes element
type -- the other use retrieve an identifier from the cache using the
string representation of the type name's corresponding special word.