fix(sem): restore error on too large discriminants
Summary
=======

Fix no error being reported for record-case discriminators that have
cover a too large value range.

Details
=======

A report was already created but it wasn't reported.

Instead of assigning the `SemReport` object to a local first, it is
passed directly to `localReport`. For visual consistency, the above
`if` branch is changed too.