spec: reorganize `basics` language specs
## Summary

Reorganize the `basics` language specification tests following the tone
of the `bootstrap` language specification tests. Tests are largely the
same but are now logically ordered based on a basic compiler
implementation.

## Details

Mostly files are reordered to follow a logical progression of
implementing a basic compiler or getting up to speed on language basics.
Readme files were updated where appropriate to reflect the new style or
indicate the need for further updates.

Some minor stylistic changes were made to rely less on yet to be tested
language features, and the replacement of references to Nim to Nimskull.