fix(sem): missed `discard` in `try/finally` errors
## Summary

A missing `discard` in a `try/finally` no longer crashes the compiler
and emits an error as it should.

## Details

The `discardCheck` used to ensure expressions are used or `void` did not
traverse `try` appropriately. Instead of looking at the last expression
in the last `except` or `try` branch, it would look at the last branch,
even if that was a `finally`. This led to incomplete data being sent for
error reporting, which ultimately led to an NPE.

Along with the fix a regression test has been added.