fix(sem): restore error on too large discriminants (#994)
## Summary

Fix no error being reported for record-case discriminators that
cover a too large value range.

## Details

The problem was already detected and a report created, but the
report object wasn't reported.

Instead of assigning the `SemReport` object to a local first, it is
passed directly to `localReport`. For visual consistency, the above
`if` branch is changed too.