vmtypegen: implement closure -> tuple lowering
A closure is lowered into a `(proc, RootRef)` tuple. For now, the
`RootRef` type has to be set up from the outside, but once types
are lowered during the MIR phase, the type lowering can be
in `vmtypegen` can be removed again.

Whether a `tyProc` type represents a closure or procedure type
is context dependent, so `genType` together with all related
type-creation routines now have a `noClosure` parameter. This
also means that `tyProc` types must not be cached.