refactor(sem): remove duplicated type assignment (#1000)
## Summary

De-duplicate the handling of assigning a symbol's type by re-using the
unused `setVarType`.

## Details

Both `semNormalizedLetOrVar` and `semNormalizedConst` handle the type
assignment for the symbol in the same way, so it makes sense to move
the implementation into a shared procedure. The unused `setVarType`
procedure is repurposed for this and renamed to `setSymType`.