fix wrong C code for `addr` of up-conversions
Summary
=======

Fix taking the address of an up-conversion of a `ref` (either
explicitly through the use of `addr` or implicitly as part of iterator
inlining) resulting in illegal C code being generated in some cases.

Details
=======

Emitting `&x->Sup` for up-conversions of `ref` values is wrong when the
expression needs to have its address taken, so the code generator
emits a cast of the `&x` instead.

Instead of basing the emission on whether the expression needs to
support having its address taken, an analysis (`compat.isLValue`) was
performed on the conversion's operand -- if the operand was deemed as
being an l-value, the `&x` conversion code was emitted.

Since the analysis is only an approximation, this resulted in incorrect
C code being generated for some operands (e.g., `let` globals and
`lent` dereferences).

The issue is fixed by marking the temporary `TLoc` with the new
`lfPreferAddr` flag, which communicates that the returned expression is
expected to have its address taken. `upConv` and `downConv` use this
information to replace the `isLValue` heuristic.

Finally, the now obsolete `isLValue` is removed and a test for taking
the address of up-conversions is added.