fix wrong C code for `addr` of `ref` conversions (#1004)
## Summary

Fix taking the address of a conversion of a `ref` (either explicitly
through the use of `addr` or implicitly as part of iterator inlining)
resulting in illegal C code being generated in some cases.

## Details

Emitting `&x->Sup` for up-conversions of `ref` values is wrong when the
expression needs to have its address taken, so the code generator
emits a cast of the `&x` instead.

Whether to the use this form was based on the operand expression, not
on how the resulting expression is used, and since the used analysis
(`compat.isLvalue`) is itself flawed, the wrong from was sometimes
used (e.g., `let` globals and `lent` dereferences).

### Fixing the issue

What code to generated for `ref` conversion is now decided based on
whether the expression is required to be an l-value, with the callsite
communicating this information this to `upConv` and `downConv` via the
new `lfWantsLvalue` flag.

The places where the flag needs to be set is where an expression has
its address taken directly or is used as the destination of an
assignment.

### Follow-on issues

The now-proper handling surfaced two pre-existing issues:
* `cgmeth.genConv` generated `(ObjUpConv (HiddenAddr x))` AST instead
  of the correct `(HiddenAddr (ObjUpConv x))` AST
* `cgen.downConv` treated generic object instance conversions as ref-
  conversion due to missing a `skipTypes` call

Both pre-existed issues are also fixed.

### Misc

The C code generated for up/down ref conversions is slightly improved:
`addrLoc` is now used for the operand and the destination loc is marked
as indirect. This prevents a redundant deref/take-address pair in some
cases.

In addition, discarding redundant up-conversion are now already skipped
in `mirgen`.

---------

Co-authored-by: Saem Ghani <saemghani+github@gmail.com>