integrate the range check into the test
Switch to a different approach: instead of moving the optimization from
`cgen` into `mirgen`, simply integrate the range check into the test
itself. This is much simpler and it also means that there's no complex
set of rules to remember (or be warned of).

Removing the warning is up next.