fix a destruction bug for global-backed locals
Summary
=======

Fix stale state in some cases reaching destructors for locals defined
in loops that are themselves not part of a routine.

Fixes https://github.com/nim-works/nimskull/issues/1010.

Details
=======

Locations defined within loops outside of routines are globals
internally. Globals so far didn't use the default-initialized-at-scope-
entry behaviour that routine-level locals use, which caused their
destructors observing them in a not-reset state in certain situations.

For ensuring proper default-initialization:
1. the pass for removing 'def's of nested globals is removed. This also
   speeds up the MIR phase
2. removing the 'def's of globals is performed during the MIR ->
   `CgNode` translation. This allows for `cgirgen` to emit a default-
   initialization for globals that need it

The facility already used for locals (i.e. moving the definition when
it's not placed immediately within a MIR scope) is also used for
globals.

This does **not** fix the fundamental issue of cleanup also being
applied to not-yet-alive globals, but it makes the behaviour between
globals and locals consistent.